# src/Core/GameLending.Core.Api/openapi.yaml
openapi: 3.1.1
info:
  title: GameLending Core
  description: Gerenciamento de empréstimos de jogos.
  version: v1
paths:
...
  /api/v1/loans:
    get:
      tags:
      - Loans
      description: 'Permissões exigidas (scope e role): loans.read'
      parameters:
      - name: friendId
        in: query
        schema:
          type: string
          format: uuid
      - name: gameId
        in: query
        schema:
          type: string
          format: uuid
      - name: gameIds
        in: query
        schema:
          type: array
          items:
            type: string
            format: uuid
      - name: page
        in: query
        schema:
          type: integer
          format: int32
          default: 1
      - name: pageSize
        in: query
        schema:
          type: integer
          format: int32
          default: 20
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/LoanViewPage'
            application/json:
              schema:
                $ref: '#/components/schemas/LoanViewPage'
            text/json:
              schema:
                $ref: '#/components/schemas/LoanViewPage'
        '400':
          description: Entrada inválida
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '401':
          description: Token ausente ou inválido
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '403':
          description: Permissão insuficiente
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '404':
          description: Recurso não encontrado
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '409':
          description: Conflito de negócio
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '500':
          description: Erro interno
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
    post:
      tags:
      - Loans
      description: 'Permissões exigidas (scope e role): loans.write'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateLoan'
          text/json:
            schema:
              $ref: '#/components/schemas/CreateLoan'
          application/*+json:
            schema:
              $ref: '#/components/schemas/CreateLoan'
      responses:
        '201':
          description: Created
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/LoanView'
            application/json:
              schema:
                $ref: '#/components/schemas/LoanView'
            text/json:
              schema:
                $ref: '#/components/schemas/LoanView'
        '400':
          description: Entrada inválida
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '401':
          description: Token ausente ou inválido
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '403':
          description: Permissão insuficiente
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '404':
          description: Recurso não encontrado
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '409':
          description: Conflito de negócio
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '500':
          description: Erro interno
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  /api/v1/loans/active:
    get:
      tags:
      - Loans
      description: 'Permissões exigidas (scope e role): loans.read'
      parameters:
      - name: friendId
        in: query
        schema:
          type: string
          format: uuid
      - name: gameId
        in: query
        schema:
          type: string
          format: uuid
      - name: gameIds
        in: query
        schema:
          type: array
          items:
            type: string
            format: uuid
      - name: page
        in: query
        schema:
          type: integer
          format: int32
          default: 1
      - name: pageSize
        in: query
        schema:
          type: integer
          format: int32
          default: 20
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/LoanViewPage'
            application/json:
              schema:
                $ref: '#/components/schemas/LoanViewPage'
            text/json:
              schema:
                $ref: '#/components/schemas/LoanViewPage'
        '400':
          description: Entrada inválida
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '401':
          description: Token ausente ou inválido
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '403':
          description: Permissão insuficiente
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '404':
          description: Recurso não encontrado
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '409':
          description: Conflito de negócio
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '500':
          description: Erro interno
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  /api/v1/loans/{id}:
    get:
      tags:
      - Loans
      description: 'Permissões exigidas (scope e role): loans.read'
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/LoanView'
            application/json:
              schema:
                $ref: '#/components/schemas/LoanView'
            text/json:
              schema:
                $ref: '#/components/schemas/LoanView'
        '400':
          description: Entrada inválida
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '401':
          description: Token ausente ou inválido
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '403':
          description: Permissão insuficiente
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '404':
          description: Recurso não encontrado
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '409':
          description: Conflito de negócio
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '500':
          description: Erro interno
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  /api/v1/loans/{id}/return:
    post:
      tags:
      - Loans
      description: 'Permissões exigidas (scope e role): loans.write'
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/LoanView'
            application/json:
              schema:
                $ref: '#/components/schemas/LoanView'
            text/json:
              schema:
                $ref: '#/components/schemas/LoanView'
        '400':
          description: Entrada inválida
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '401':
          description: Token ausente ou inválido
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '403':
          description: Permissão insuficiente
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '404':
          description: Recurso não encontrado
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '409':
          description: Conflito de negócio
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '500':
          description: Erro interno
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  /api/v1/statistics:
    get:
      tags:
      - Statistics
      description: 'Permissões exigidas (scope e role): games.read, friends.read,
        loans.read'
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/Statistics'
            application/json:
              schema:
                $ref: '#/components/schemas/Statistics'
            text/json:
              schema:
                $ref: '#/components/schemas/Statistics'
        '400':
          description: Entrada inválida
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '401':
          description: Token ausente ou inválido
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '403':
          description: Permissão insuficiente
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '404':
          description: Recurso não encontrado
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '409':
          description: Conflito de negócio
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '500':
          description: Erro interno
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
...
components:
  schemas:
    CreateLoan:
      type: object
      properties:
        gameId:
          type: string
          format: uuid
        friendId:
          type: string
          format: uuid
      additionalProperties: false
...
    LoanView:
      type: object
      properties:
        id:
          type: string
          format: uuid
        gameId:
          type: string
          format: uuid
        gameTitle:
          type:
          - 'null'
          - string
        friendId:
          type: string
          format: uuid
        friendName:
          type:
          - 'null'
          - string
        loanedAt:
          type: string
          format: date-time
        returnedAt:
          type:
          - 'null'
          - string
          format: date-time
      additionalProperties: false
    LoanViewPage:
      type: object
      properties:
        items:
          type:
          - 'null'
          - array
          items:
            $ref: '#/components/schemas/LoanView'
        total:
          type: integer
          format: int32
        pageNumber:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
      additionalProperties: false
    ProblemDetails:
      type: object
      properties:
        type:
          type:
          - 'null'
          - string
        title:
          type:
          - 'null'
          - string
        status:
          type:
          - 'null'
          - integer
          format: int32
        detail:
          type:
          - 'null'
          - string
        instance:
          type:
          - 'null'
          - string
      additionalProperties: {}
...
    Statistics:
      type: object
      properties:
        totalGames:
          type: integer
          format: int32
        availableGames:
          type: integer
          format: int32
        borrowedGames:
          type: integer
          format: int32
        totalFriends:
          type: integer
          format: int32
        activeLoans:
          type: integer
          format: int32
      additionalProperties: false
  securitySchemes:
    Bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
- Bearer: []
...